% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visit_detail.R
\name{column_jump}
\alias{column_jump}
\title{collapse_over}
\usage{
column_jump(dt, col, order_vars, group, col_jump)
}
\arguments{
\item{dt}{data.table}

\item{col}{output column}

\item{order_vars}{vars for sorting}

\item{group}{grouping variable}

\item{col_jump}{column where discontinuity will be flagged}
}
\description{
Wrangles a time series of (bed) moves and identifies department transfers etc.
}
