% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_any}
\alias{round_any}
\title{utils}
\usage{
round_any(x, accuracy = 1)
}
\arguments{
\item{ctn}{database connection}

\item{target_schema}{target schema}

\item{table_name}{table name}

\item{returnDT}{return as data.table instead of data.frame}
}
\description{
utility functions for working with star
}
