% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_star_from}
\alias{select_star_from}
\title{utils}
\usage{
select_star_from(ctn, target_schema, table_name, returnDT = TRUE)
}
\arguments{
\item{ctn}{database connection}

\item{target_schema}{target schema}

\item{table_name}{table name}

\item{returnDT}{return as data.table instead of data.frame}
}
\description{
utility functions for working with star
}
