% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{parse_coalesce_functions}
\alias{parse_coalesce_functions}
\title{return a character vector of function names
expects either alist (or) a character vector}
\usage{
parse_coalesce_functions(funs = c("first"))
}
\description{
return a character vector of function names
expects either alist (or) a character vector
}
